### 2023-10-01 #5914 @editorialbot
[PRE REVIEW]: NBI: A library for Nystrom Boundary Integral calculations
**Submitting author:** @mjcarley (<a href="http://orcid.org/0000-0003-2965-8984">Michael Carley</a>)
**Repository:** https://github.com/mjcarley/nbi/
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
