### 2023-10-01 #5914 @danielskatz
@mjcarley - thanks for your submission to JOSS.

Before we can start a review, we have some issues to resolve.

1. You could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.

2. There are an awful lot of files in the top level of the git repository, including what appear to be log files. It seems like this might be cleaned up a bit. Also, the testing instructions need to be more explicit, I think. And the README has some formatting issues, where what seem like they should be headers are instead bullets.
