### 2023-10-02 28c7dc3
added dois to paper.bib; removed log files
