+++ JOSS/paper.bib
@@ -31,7 +31,7 @@ volume = {10},
 pages = {100092},
 year = {2021},
 issn = {2590-0552},
-url = {https://doi.org/10.1016/j.jcpx.2021.100092}
+doi = {10.1016/j.jcpx.2021.100092}
 }
 
 @Book{gumerov-duraiswami04,
@@ -40,7 +40,9 @@ url = {https://doi.org/10.1016/j.jcpx.2021.100092}
                   in three dimensions},
   publisher = 	 {Elsevier},
   year = 	 2004,
-  address = 	 {Oxford}}
+  address = 	 {Oxford},
+  doi =          {10.1016/b978-0-08-044371-3.x5000-5}
+  }
 
 @TechReport{gumerov-duraiswami05,
   author = 	 {Nail A. Gumerov and Ramani Duraiswami},
@@ -81,7 +83,7 @@ url = {https://doi.org/10.1016/j.jcpx.2021.100092}
                   boundary element library}, 
   journal = 	 {Journal of Open Source Software},
   year = 	 2021,
-  url =          {https://joss.theoj.org/papers/10.21105/joss.02879},
+  doi =          {10.21105/joss.02879},
   volume = 	 6,
   number = 	 59}
 
