### 2023-10-02 84f5170
doc: added heading formatting to README.md
