### 2023-10-02 ad96775
doc: added heading formatting to examples/README.md
