### 2023-10-02 20c6243
doc: minor formatting READMEs
