### 2023-10-02 2e9fc14
doc: expanded examples/README.md
