### 2023-10-02 b4e951d
doc: reformatting examples list in examples/README.md
