+++ examples/README.md
@@ -14,26 +14,26 @@ interior source. The NBI solver returns error estimates (L_2, L_inf,
 and r.m.s.) computed as the difference between the potential specified
 in the boundary condition file, and that computed by the solver.
 
-The output from the calculation which be visualized using GMSH is two
-`.msh` files, one named according to the geometry of the problem,
+The output from the calculation which can be visualized using GMSH is
+two `.msh` files, one named according to the geometry of the problem,
 e.g. `sphere.msh`, and one called `grid.msh` which shows the computed
-field around the surface. 
+field around the surface.
 
-To run a test case, from inside the directory, use the script
+To run a test case, from inside the directory use the script
 `solve-helmholtz` or `solve-laplace` as appropriate. For example:
 
 ```
-cd SphereHelmholtz
+cd .../examples/SphereHelmholtz
 ../solve-helmholtz -k 1.0
-gmsh sphere.msh`
+gmsh sphere.msh
 ```
 
 generates a spheroidal surface and solves the boundary integral
 problem with an interior point source of wavenumber 1.0. The command
-`gmsh sphere.msh` displays the solution surface potential. The output
-from the solver will be in a file called `solution.log`, which
-contains details of the progress of the solution. The last few lines
-will look something like this:
+`gmsh sphere.msh` displays the solution surface potential. Output from
+the solver will be in a file called `solution.log`, which contains
+details of the progress of the solution. The last few lines will look
+something like this:
 
 ```
 nbi-solve-helmholtz: 1 iterations; error = 5.09449e-07 [287.509] (287.48)
@@ -62,28 +62,33 @@ and
 
 These will enter each directory in turn and run the tests, logging
 output to `test-[laplace|helmholtz].log` and estimated error norms to
-`test-[laplace|helmholtz]-norm.log`
+`test-[laplace|helmholtz]-norm.log`.
 
-SphereLaplace: solve for the Laplace potential on an ellipsoid
-	       (deformed sphere)
+# Test cases
 
-SphereHelmholtz: solve for the Helmholtz (wave equation) potential on
-		 an ellipsoid under excitation from a sinusoidally varying
-		 ring source (this can be used to simulate rotor and
-		 propeller noise).
+SphereLaplace
+: solve for the Laplace potential on an ellipsoid (deformed sphere)
 
-Flow: solve the Laplace equation to compute the potential around a
+SphereHelmholtz
+: solve for the Helmholtz (wave equation) potential on an ellipsoid
+		 under excitation from a sinusoidally varying ring source
+		 (this can be used to simulate rotor and propeller noise).
+
+Flow
+: solve the Laplace equation to compute the potential around a
 	  sphere in a uniform flow. The error norms reported by the solver
 	  in this case are not meaningful, because the boundary condition
 	  is not generated by an internal source. Instead, the accuracy of
 	  the result can be checked in Octave, using a script in the
-	  directory. The README file in the directory explains how this
-	  can be checked using Octave.
+	  directory. The README file in the directory explains how to do
+	  this. 
 
-Wing: solve the Laplace or Helmholtz problem for a point source inside
+Wing
+: solve the Laplace or Helmholtz problem for a point source inside
       a wing-type geometry, generated using the AGG library. 
 
-Scattering: solve for scattering of a plane wave by a sphere, with
+Scattering
+: solve for scattering of a plane wave by a sphere, with
 	    comparison to the analytical series solution. The error norms
 	    reported by the solver in this case are not meaningful,
 	    because the boundary condition is not generated by an internal
@@ -92,11 +97,13 @@ Scattering: solve for scattering of a plane wave by a sphere, with
 	    the directory contains instructions on how to perform the
 	    check.
 
[…]
