### 2023-10-02 d05c0f5
fact: removing files which should be ignored by .gitignore
