+++ doc/latex/group__boundary.tex
@@ -1,226 +1,0 @@
-\doxysection{Boundary conditions}
-\label{group__boundary}\index{Boundary conditions@{Boundary conditions}}
-
-
-Definition and evaluation of boundary conditions.  
-
-
-\doxysubsection*{Functions}
-\begin{DoxyCompactItemize}
-\item 
-gint \textbf{ nbi\+\_\+boundary\+\_\+condition\+\_\+set} (\textbf{ nbi\+\_\+surface\+\_\+t} $\ast$s, gdouble $\ast$p, gint pstr, gdouble $\ast$pn, gint nstr, \textbf{ nbi\+\_\+boundary\+\_\+condition\+\_\+t} $\ast$b)
-\begin{DoxyCompactList}\small\item\em compute boundary condition on surface \end{DoxyCompactList}\end{DoxyCompactItemize}
-\begin{DoxyCompactItemize}
-\item 
-gint \textbf{ nbi\+\_\+functions\+\_\+list} (FILE $\ast$f, gboolean help)
-\item 
-\textbf{ nbi\+\_\+boundary\+\_\+condition\+\_\+t} $\ast$ \textbf{ nbi\+\_\+boundary\+\_\+condition\+\_\+new} (\textbf{ nbi\+\_\+problem\+\_\+t} problem)
-\item 
-gint \textbf{ nbi\+\_\+boundary\+\_\+condition\+\_\+read} (FILE $\ast$f, \textbf{ nbi\+\_\+boundary\+\_\+condition\+\_\+t} $\ast$b)
-\item 
-gint \textbf{ nbi\+\_\+boundary\+\_\+condition\+\_\+has\+\_\+variable} (\textbf{ nbi\+\_\+boundary\+\_\+condition\+\_\+t} $\ast$b, gchar $\ast$v)
-\item 
-gint \textbf{ nbi\+\_\+boundary\+\_\+condition\+\_\+write} (FILE $\ast$f, \textbf{ nbi\+\_\+boundary\+\_\+condition\+\_\+t} $\ast$b)
-\item 
-gint \textbf{ nbi\+\_\+boundary\+\_\+condition\+\_\+eval} (\textbf{ nbi\+\_\+boundary\+\_\+condition\+\_\+t} $\ast$b, gdouble $\ast$x, gdouble $\ast$n)
-\item 
-gboolean \textbf{ nbi\+\_\+boundary\+\_\+condition\+\_\+defined} (\textbf{ nbi\+\_\+boundary\+\_\+condition\+\_\+t} $\ast$b, gchar $\ast$v)
-\end{DoxyCompactItemize}
-
-
-\doxysubsection{Detailed Description}
-Definition and evaluation of boundary conditions. 
-
-NBI solves boundary integral problems subject to a Neumann boundary condition, which corresponds to a surface normal velocity. This is the boundary condition appropriate to acoustic scattering from a hard surface, or to solving a potential problem in fluid dynamics. The boundary conditions are defined using analytical expressions, which are evaluated at each surface node using the {\texttt{ tinyexpr}} parser which accepts standard mathematical expressions, and a number of built-\/in functions for standard cases, such as the point sources in the Laplace and Helmholtz problems. The executables {\ttfamily nbi-\/solve-\/laplace} and {\ttfamily nbi-\/solve-\/helmholtz} accept the {\ttfamily -\/B} option which lists the built-\/in functions which can be used to define boundary conditions.
-
[…]
