+++ doc/latex/group__matrix.tex
@@ -1,683 +1,0 @@
-\doxysection{Matrix operations}
-\label{group__matrix}\index{Matrix operations@{Matrix operations}}
-
-
-Matrix operations and iterative solvers.  
-
-
-\doxysubsection*{Functions}
-\begin{DoxyCompactItemize}
-\item 
-\textbf{ nbi\+\_\+matrix\+\_\+t} $\ast$ \textbf{ nbi\+\_\+matrix\+\_\+new} (\textbf{ nbi\+\_\+surface\+\_\+t} $\ast$s)
-\begin{DoxyCompactList}\small\item\em generate a new \doxyref{nbi\+\_\+matrix\+\_\+t}{p.}{group__matrix_gae0e1ce93ef5eb67f95f90bb28069d3c4} for use in solvers \end{DoxyCompactList}\item 
-gint \textbf{ nbi\+\_\+matrix\+\_\+read} (FILE $\ast$f, \textbf{ nbi\+\_\+matrix\+\_\+t} $\ast$m)
-\begin{DoxyCompactList}\small\item\em read an \doxyref{nbi\+\_\+matrix\+\_\+t}{p.}{group__matrix_gae0e1ce93ef5eb67f95f90bb28069d3c4} from file \end{DoxyCompactList}\item 
-gint \textbf{ nbi\+\_\+matrix\+\_\+write} (FILE $\ast$f, \textbf{ nbi\+\_\+matrix\+\_\+t} $\ast$m)
-\begin{DoxyCompactList}\small\item\em write an \doxyref{nbi\+\_\+matrix\+\_\+t}{p.}{group__matrix_gae0e1ce93ef5eb67f95f90bb28069d3c4} to file \end{DoxyCompactList}\item 
-gint \textbf{ nbi\+\_\+matrix\+\_\+fmm\+\_\+init} (\textbf{ nbi\+\_\+matrix\+\_\+t} $\ast$m, \textbf{ nbi\+\_\+problem\+\_\+t} problem, wbfmm\+\_\+shift\+\_\+operators\+\_\+t $\ast$shifts, guint $\ast$order\+\_\+s, gint sstr, guint $\ast$order\+\_\+r, gint rstr, gint depth, gdouble dtree, gboolean shift\+\_\+bw, gboolean precompute\+\_\+local, gdouble $\ast$work)
-\begin{DoxyCompactList}\small\item\em initialize Fast Multipole Method matrix \end{DoxyCompactList}\item 
-gint \textbf{ nbi\+\_\+matrix\+\_\+multiply} (\textbf{ nbi\+\_\+matrix\+\_\+t} $\ast$A, gdouble $\ast$x, gint xstr, gdouble al, gdouble $\ast$y, gint ystr, gdouble bt, gint nthreads, gdouble $\ast$work)
-\begin{DoxyCompactList}\small\item\em perform matrix multiplication $y := \beta y + \alpha A x$ \end{DoxyCompactList}\item 
-\textbf{ nbi\+\_\+matrix\+\_\+t} $\ast$ \textbf{ nbi\+\_\+matrix\+\_\+assemble\+\_\+helmholtz} (\textbf{ nbi\+\_\+surface\+\_\+t} $\ast$s, gdouble k, gdouble eta, gint nqa, gint dmax, gdouble tol, gint N, gint nu, gint nnmax, gint nthreads)
-\item 
[…]
