+++ doc/latex/modules.tex
@@ -1,7 +1,0 @@
-\doxysection{Modules}
-Here is a list of all modules\+:\begin{DoxyCompactList}
-\item \contentsline{section}{Surfaces}{\pageref{group__surfaces}}{}
-\item \contentsline{section}{Matrix operations}{\pageref{group__matrix}}{}
-\item \contentsline{section}{Boundary conditions}{\pageref{group__boundary}}{}
-\item \contentsline{section}{Export and visualisation}{\pageref{group__export}}{}
-\end{DoxyCompactList}

[…]
