### 2023-10-02 829b97d
fact: removing files which should be ignored by .gitignore
