### 2023-10-02 #5914 @mjcarley
@danielskatz thank you for the comments. I have:

added the missing DOIs;

removed many of files at the top level and updated .gitignore;

updated the testing instructions in README.md in the examples directory;

reformatted the README.md to add headings where they are more appropriate.
