### 2023-11-26 mjcarley/nbi#1 @bonh
license file
Perhaps you want to replace COPYING with LICENSE?

JOSS specifically asks "Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?".

But COPYING is standard for gnu gpl projects? And github recognizes the COPYING file, too.

So I don't know.

 openjournals/joss-reviews/issues/5972
