### 2023-11-26 #5972 @bonh
@vissarion The repo does not contain a LICENSE file but a COPYING file. The checklist specifically asks for the former but github seems to be okay with the latter, too. Please check the issue.
