### 2023-11-27 mjcarley/nbi#1 @vissarion
Copying from https://joss.readthedocs.io/en/latest/review_criteria.html

> There should be an [OSI approved](https://opensource.org/licenses/alphabetical) license included in the repository. >Common licenses such as those listed on [choosealicense.com](https://choosealicense.com/) are preferred. Note there >should be an actual license file present in the repository not just a reference to the license.
>
> Acceptable: A plain-text LICENSE or COPYING file with the contents of an OSI approved license"

So I think we are OK.

@bonh please edit the review checklist point from
"Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?".
to
"Does the repository contain a plain-text LICENSE or COPYING file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?"

to be more clear.
