### 2023-12-01 mjcarley/nbi#2 @bonh
JOSS: Comparison to code cited in Greengard et al. 2021
You cite Greengard et al. 2021 as the basis for your implementation ("NBI implements recent work on high-order
24 solution of boundary integral problems (Greengard et al., 2021) in a form which can be used
25 for a range of engineering applications."). It seems that they provide a reference implementation of their paper at https://github.com/fastalgorithms/fmm3dbie. I have not tested their code or your code, yet.

Perhaps you should make clear how their code and yours differ?

https://github.com/openjournals/joss-reviews/issues/5972
