### 2023-12-01 mjcarley/nbi#3 @bonh
JOSS: mention your paper?
I was searching the web looking for papers referencing your library. While I did not find any (perhaps not surprising, your code is still young, but let me know!) I found your arxiv paper https://arxiv.org/pdf/2301.01179.pdf. It is somewhat different, but perhaps you have a different paper in the pipeline which makes use of your library?
