### 2023-12-01 mjcarley/nbi#2 @bonh
That publication by the same sphere of people might be relevant, too.

https://epubs.siam.org/doi/epdf/10.1137/22M1514040
