### 2023-12-01 #5972 @bonh
Right now I'm assessing the "Substantial scholarly effort" as requested by guidelines:

- Age of software (is this a well-established software project) / length of commit history. _not a problem, only a few months old, right?_
- Number of commits. _a lot of work has been done on the code_
- Number of authors. _only one, but that's alright, I guess_
- Total lines of code (LOC). _a lot of code_
- Whether the software has already been cited in academic papers. _no, right? see https://github.com/mjcarley/nbi/issues/3_
- Whether the software is sufficiently useful that it is likely to be cited by your peer group. _depends what the difference is to https://github.com/fastalgorithms/fmm3dbie, see https://github.com/mjcarley/nbi/issues/2_
