### 2023-12-01 mjcarley/nbi#3 @mjcarley
That paper describes a method for the Laplace equation only, i.e. not including the Helmholtz equation, and only in a cylindrical domain. There is not yet a paper in preparation using the current code, but I expect to use it in applications related to an ongoing project in the future.
