### 2023-12-01 844144c
doc: updated paper.md
