### 2023-12-01 mjcarley/nbi#2 @mjcarley
I have modified paper.md to include a brief statement on how the code relates to Greengard et al:

The intention has been to develop an implementation of the main
elements of [@greengard-oneil-rachh-vico21] in a code which is
relatively easy to use in applications. The principal differences
which a user will note between NBI and the reference implementation of
the method of [@greengard-oneil-rachh-vico21] are: the range of
geometry formats supported, including the widely used GMSH standard;
the use of the standard PETSc iterative solvers; and the ability to
define boundary conditions using a built-in parser for mathematical
expressions.
