### 2023-12-04 mjcarley/nbi#5 @bonh
JOSS: your dependecies
You list four of your packages as dependencies.

Perhaps you'd like to package them into `nbi` via `git-submodule` or `git-subtree` or ... (there are a bunch of options, don't know which one is best), and automatize the installation?
