### 2023-12-04 #5972 @bonh
I have trouble compiling the package, see https://github.com/mjcarley/sqt/issues/1.
