### 2023-12-04 d3f820b
doc: updated README.md to reflect dependencies in submodules
