### 2023-12-04 #5972 @mjcarley
I have included the four dependencies mentioned as submodules and updated README.md to include a note on using git clone --recursive when downloading the code.
