### 2023-12-04 832d456
doc: updated README.md to reflect dependencies in submodules
