### 2023-12-04 948dbb1
doc: updated README.md to reflect dependencies in submodules
