### 2023-12-04 cc20be5
doc: updated README.md to reflect dependencies in submodules
