### 2023-12-04 4d2b50a
bug: autogen.sh adds m4 directory
