+++ autogen.sh
@@ -1,5 +1,7 @@
 #! /bin/sh
 
+mkdir -p m4
+
 aclocal -I . \
 && autoheader \
 && automake --add-missing \
