### 2023-12-06 mjcarley/nbi#5 @bonh
Pasted @mjcarley's comment from https://github.com/openjournals/joss-reviews/issues/5972#issuecomment-1838947294

> I have included the four dependencies mentioned as submodules and updated README.md to include a note on using git clone --recursive when downloading the code.
