### 2023-12-06 33d6dce
maint: updated configure.ac to use PKG_CHECK_MODULES for glib
