### 2023-12-06 cb08f0b
maint: updated configure.ac to use PKG_CHECK_MODULES for glib
