### 2023-12-07 #5972 @bonh
To optimize my time effort I will pause my review until @mjcarley has decided on https://github.com/mjcarley/nbi/issues/5 :smile:.
