+++ src/documentation.c
@@ -1,6 +1,6 @@
 /*
  * Autogenerated file, do not  edit
- * Wed Dec  6 14:32:29 GMT 2023
+ * Thu Dec  7 18:47:41 GMT 2023
  * -------------------------------
  */
 
