+++ config.h
@@ -25,7 +25,7 @@
 #define HAVE_AVX_INSTRUCTIONS 1
 
 /* BLAS wrapper header available */
-/* #undef HAVE_BLASWRAP */
+#define HAVE_BLASWRAP 1
 
 /* Define to 1 if you have the <dlfcn.h> header file. */
 #define HAVE_DLFCN_H 1
