+++ src/documentation.c
@@ -1,6 +1,6 @@
 /*
  * Autogenerated file, do not  edit
- * Thu Dec  7 18:47:41 GMT 2023
+ * Thu Dec  7 19:26:54 GMT 2023
  * -------------------------------
  */
 
