+++ src/export.c
@@ -14,6 +14,10 @@
  * along with NBI.  If not, see <https://www.gnu.org/licenses/>.
  */
 
+#ifdef HAVE_CONFIG_H
+#include <config.h>
+#endif /*HAVE_CONFIG_H*/
+
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -24,12 +28,13 @@
 
 #include <nbi.h>
 
+#ifdef HAVE_SQT
 #include <sqt.h>
-#include <blaswrap.h>
+#endif /*HAVE_SQT*/
 
-#ifdef HAVE_CONFIG_H
-#include <config.h>
-#endif /*HAVE_CONFIG_H*/
+#ifdef HAVE_BLASWRAP
+#include <blaswrap.h>
+#endif /*HAVE_BLASWRAP*/
 
 #include "nbi-private.h"
 
