+++ tools/nbi-process.c
@@ -14,7 +14,6 @@
  * along with NBI.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-
 #ifdef HAVE_CONFIG_H
 #include <config.h>
 #endif /*HAVE_CONFIG_H*/
@@ -29,10 +28,17 @@
 
 #include <nbi.h>
 
+#ifdef HAVE_WBFMM
 #include <wbfmm.h>
-#include <sqt.h>
+#endif /*HAVE_WBFMM*/
 
+#ifdef HAVE_SQT
+#include <sqt.h>
+#endif /*HAVE_SQT*/
+
+#ifdef HAVE_BLASWRAP
 #include <blaswrap.h>
+#endif /*HAVE_BLASWRAP*/
 
 #include "nbi-private.h"
 
