### 2023-12-11 57f7e88
maint: updating src/Makefile.am and configure.ac
