+++ src/documentation.c
@@ -1,6 +1,6 @@
 /*
  * Autogenerated file, do not  edit
- * Sat Dec  9 17:44:15 GMT 2023
+ * Mon Dec 11 14:52:30 GMT 2023
  * -------------------------------
  */
 
