### 2023-12-11 a48f2e3
maint: updating src/Makefile.am and configure.ac
