### 2023-12-11 3cad54b
maint: modified src/Makefile.am to include external dependencies
