+++ src/documentation.c
@@ -1,6 +1,6 @@
 /*
  * Autogenerated file, do not  edit
- * Mon Dec 11 15:04:33 GMT 2023
+ * Mon Dec 11 16:18:25 GMT 2023
  * -------------------------------
  */
 
