### 2023-12-11 b327d42
maint: modified src/Makefile.am to include external dependencies
