### 2023-12-11 5c721e8
maint: symbolic links to external headers in configure.ac
