### 2023-12-11 a949d17
maint: symbolic links to external headers in configure.ac
