+++ src/agg.h
@@ -1,1 +1,0 @@
-externals/agg/src/agg.h
