+++ src/sqt.h
@@ -1,1 +1,0 @@
-externals/sqt/src/sqt.h
