+++ src/wbfmm.h
@@ -1,1 +1,0 @@
-externals/wbfmm/src/wbfmm.h
