### 2023-12-12 #5972 @vissarion
Hi @mscroggs are you still available to review this JOSS paper?
