### 2023-12-12 #5972 @vissarion
Hi @mjcarley it seems that one reviewer is waiting for your feedback in https://github.com/mjcarley/nbi/issues/5
I see that you have provided a solution to the dependencies issue but please resolve https://github.com/mjcarley/nbi/issues/5 by agreeing on a solution with @bonh so that we can move on with the review. Thanks!
