### 2023-12-12 7cc48ee
maint: updating module integration; CFLAGS for headers in submodule directories
