+++ src/documentation.c
@@ -1,6 +1,6 @@
 /*
  * Autogenerated file, do not  edit
- * Mon Dec 11 17:56:24 GMT 2023
+ * Tue Dec 12 09:38:22 GMT 2023
  * -------------------------------
  */
 
