### 2023-12-12 mjcarley/nbi#5 @mjcarley
Three of the modules are now bundled into the main repository and included in the build scripts. The exception is blaswrap, because that is a dependency for the submodules (it is no more than a wrapper for BLAS) and I did not want to have multiple versions of it in the dependencies and subdependencies.

The README.md now lists the dependencies of the submodules as well as of NBI proper to allow installation of those libraries before trying to install the submodules.
