### 2023-12-12 mjcarley/nbi#5 @bonh
Awesome! I'll try it later.

In the new `README` it says:

> To install NBI, you will need to install some or all of the following

Later, it says:

> NBI and a number of its dependencies use a wrapper to interface with BLAS, which you need to install first.

and
> You will also need the GMSH API library

That way it gets confusing what's really required. This description:
> If you are installing the AGG geometry library from within this tree, you will need to install Christian Woltering's API for Jonathan Shewchuk's Triangle code
is very helpful.

Perhaps it would be easier if you separate dependencies into requirements and added functionalities. And I prefer a nice description which package adds which functionality. :smile:
