### 2023-12-12 a28c6eb
maint: rewrite autogen.sh to include autoheader
