+++ autogen.sh
@@ -1,3 +1,8 @@
 #! /bin/sh
 
-autoreconf --install
+mkdir -p m4
+
+aclocal -I .
+autoheader
+automake --add-missing
+autoconf
