### 2023-12-12 #5972 @mscroggs
> Hi @mscroggs are you still available to review this JOSS paper?

Hi. Yes, sorry I've been very slow with this review. I'll get on it this week.
