### 2023-12-12 d60470d
maint: rewrite autogen.sh to force full autoreconf
