+++ autogen.sh
@@ -1,8 +1,3 @@
 #! /bin/sh
 
-mkdir -p m4
-
-aclocal -I .
-autoheader
-automake --add-missing
-autoconf
+autoreconf --force --install
