+++ README.md
@@ -6,44 +6,51 @@ here:
 
 https://github.com/codeplea/tinyexpr
 
-# Prerequisites
+# Getting NBI
 
-To install NBI, you will need to install some or all of the following:
-
-- BLAS wrapper https://github.com/mjcarley/blaswrap
-- Singular Quadrature for Triangles https://github.com/mjcarley/sqt
-- Wide Band Fast Multipole Method https://github.com/mjcarley/wbfmm
-- Aerodynamic Geometry Generator https://github.com/mjcarley/agg
-- GMSH API https://www.gmsh.info/
-- Triangle API https://github.com/wo80/Triangle/
-- PETSc https://petsc.org/
-
-NBI and a number of its dependencies use a wrapper to interface with
-BLAS, which you need to install first. This is available from:
-
-- https://github.com/mjcarley/blaswrap
-
-You will also need to have installed the following libraries:
-
-- https://github.com/mjcarley/sqt
-- https://github.com/mjcarley/wbfmm
-- https://github.com/mjcarley/agg
-
-If you clone the github repository using
+NBI can be downloaded from https://github.com/mjcarley/nbi, but it is
+recommended (see under Installation) to clone it using
 
 `git clone --recursive https://github.com/mjcarley/nbi`
 
-these libraries will be downloaded automatically in the `externals`
-subdirectory of `nbi`, and will be built as part of the NBI build
-process. 
+in order to facilitate installation of dependencies as submodules.
 
-You will also need the GMSH API library, available from
+# Prerequisites
+
+NBI requires the following packages:
+- BLAS wrapper https://github.com/mjcarley/blaswrap
+- SQT (Singular Quadrature for Triangles) https://github.com/mjcarley/sqt
+- WBFMM (Wide Band Fast Multipole Method) https://github.com/mjcarley/wbfmm
+- GMSH API https://www.gmsh.info/
+
+NBI can also make use of:
+- AGG (Aerodynamic Geometry Generator) https://github.com/mjcarley/agg,
+  which requires the Triangle API https://github.com/wo80/Triangle/
+- the iterative solvers in PETSc https://petsc.org/
+
+AGG provides a geometry specification interface based on the
+methods of Brenda Kulfan https://www.brendakulfan.com/research, which
+can be used to generate surfaces such as wings found in aeronautical
+applications. 
+
+It is recommended to install PETSc to provide a choice of
+high-performance iterative solvers for the boundary integral
+equations. NBI does have a built-in basic GMRES solver but PETSc
+provides access to a range of methods which are not otherwise available.
+
+# Installation
+
+You need to install the BLAS wrapper separately before installing NBI
+or its dependencies, as it is used by a number of different
+dependencies of NBI.
+
+You will also need to install the GMSH API library, available from
 
 - https://www.gmsh.info/
 
-If you are installing the AGG geometry library from within this tree,
-you will need to install Christian Woltering's API for Jonathan
-Shewchuk's Triangle code, available from:
+If you are installing the AGG geometry library you will need to
+install Christian Woltering's API for Jonathan Shewchuk's Triangle
+code, available from:
 
 - https://github.com/wo80/Triangle/
 
@@ -55,16 +62,30 @@ basic built-in iterative solver (recommended):
 and make sure the environment variable PETSC_DIR is set and visible in
 the shell.
 
-# Installation
+To install NBI, it is to clone the NBI repository using
 
-If you have downloaded the source from github or equivalent, you will
-need the autotools suite. Generate the configure script with
+`git clone --recursive https://github.com/mjcarley/nbi`
+
+This will download the NBI repository and SQT, WBFMM, and AGG as
+submodules. In the top directory of the repository, generate the
+configuration scripts using
 
[…]
