### 2024-01-12 mjcarley/nbi#5 @bonh
> Have you thought about including the dependencies in the build script of `nbi` ?

Hi @mjcarley, have you made a decision concerning the "automated package management solution"?
