### 2024-01-12 #5972 @bonh
I'm still not able to install the dependency `blaswrap` from @mjcarley, see https://github.com/mjcarley/blaswrap/issues/1.

All the projects make heavy use of `automake`, which I'm familiar with but not an expert by far. So I really hope that it's not my own stupidity.
