### 2024-01-15 mjcarley/nbi#5 @mjcarley
The additional packages, other than blaswrap, are now included in the repository and are downloaded and installed if the --recursive option is used when cloning the repository.
