### 2024-01-15 mjcarley/nbi#5 @bonh
```
./autogen.sh
./configure
...
./configure: line 10971: syntax error near unexpected token `avx,'
./configure: line 10971: `AX_GCC_X86_CPU_SUPPORTS(avx, hasavx=yes)'
```
