### 2024-01-15 #5972 @bonh
WIth the help of @mjcarley I was able to install the dependency `blaswrap`, thanks! However, `nbi` is still not running for me.

@mjcarley I'm really sorry but I feel like a beta tester here. To make sure that it's not just my own incompetence I'd appreciate if you'd confirm that someone else already tried to install and use your software. @vissarion Is that reasonable?
