### 2024-01-15 mjcarley/nbi#5 @mjcarley
This seems to be the same problem with m4 search paths. I have moved the relevant m4 macros into the m4 directory as with the other package.
