### 2024-01-15 #5972 @bonh
Good to know! :+1:

Is your system/platform that special? I tried installing it with a vanilla Ubuntu 20.

Perhaps it would help if you could use a common distribution in a virtual environment, install the dependencies via conda or so and install and run your software there?
