### 2024-01-15 #5972 @mjcarley
It's a standard up to date Slackware installation. I can get access to a different (standard) system and will try installing there out of the box.
