### 2024-01-15 #5972 @vissarion
Thanks @mjcarley that would be great. You may consider using github actions (that have some standard Ubuntu environments) This will have the advantage that we will know if the code is still installable in future changes. Of course this is not mandatory but just a suggestion.
