### 2024-01-17 23a0c76
feat: update tools/Makefile.am to deal with not having PETSc
