### 2024-01-19 0d1d127
maint: checking for PETSc version number in configure.ac
