+++ src/documentation.c
@@ -1,872 +1,0 @@
-/*
- * Autogenerated file, do not  edit
- * Tue Dec 12 09:38:22 GMT 2023
- * -------------------------------
- */
-
-
-/**
- * @ingroup surfaces
- *
- * @brief allocate an ::nbi_surface_t
- *
- * @param nnmax maximum number of nodes in surface;
- * @param npmax maximum number of patches in surface.
- *
- * @return pointer to newly allocated ::nbi_surface_t
- **/
-
-nbi_surface_t *nbi_surface_alloc(gint nnmax, gint npmax) ;
-
-/**
- * @ingroup surfaces
- *
- * @brief write an ::nbi_surface_t to file
- *
- * @param s surface to write to file
- * @param f output file stream
- *
- * @return 0 on success
- **/
-
-gint nbi_surface_write(nbi_surface_t *s, FILE *f) ;
-
-/**
- * @ingroup surfaces
- *
- * @brief allocate and read an ::nbi_surface_t from file
- *
- * @param f input file stream.
- *
- * @return pointer to newly allocated surface, read from file
- **/
-
-nbi_surface_t *nbi_surface_read(FILE *f) ;
-
-/**
- * @ingroup surfaces
- *
- * @brief estimate centroid of surface patch
- *
- * @param x coordinates of patch nodes;
- * @param xstr stride from first component of one to first component of next;
- * @param w patch quadrature weights;
- * @param wstr quadrature weight stride;
- * @param nx number of nodes;
- * @param c on exit contains coordinates of centroid.
- *
- * @return 0 on success.
- **/
-
-gint nbi_surface_patch_centroid(gdouble *x, gint xstr,
-				      gdouble *w, gint wstr,
-				      gint nx,
-				      gdouble *c) ;
-
-/**
- * @ingroup surfaces
- *
- * @brief estimate radius of circumsphere of surface patch 
- *
- * @param x coordinates of patch nodes;
- * @param xstr stride from first component of one node to first component 
- * of next;
- * @param nx number of patch nodes;
- * @param c centroid of patch (from 
- * nbi_surface_patch_centroid)).
- *
- * @return patch radius
- **/
-
-gdouble nbi_surface_patch_radius(gdouble *x, gint xstr,
-					gint nx, gdouble *c) ;
-
-/**
- * @ingroup surfaces
- *
- * @brief identify neighbour nodes of surface patch
- *
- * @param c centroid;
- * @param r radius of neighbourhood;
- * @param x list of candidate neighbour nodes;
- * @param xstr stride from first component of one node to first component 
- * of next;
- * @param nx number of candidate neighbour nodes;
- * @param n0 first entry in \a x to check;
- * @param n1 last entry in \a x to check;
- * @param nbrs on exit contains indices of nodes lying within distance 
- * \a r of \a c; 
- * @param nnbrs on exit contains number of nodes in \a nbrs;
- * @param nnmax maximum number of neighbours to be inserted in \a nbrs.
- *
- * @return 0 on success.
- **/
-
-gint nbi_patch_neighbours(gdouble *c, gdouble r,
-				gdouble *x, gint xstr, gint nx,
-				gint n0, gint n1,
-				gint *nbrs, gint *nnbrs,
-				gint nnmax) ;
-
-/**
- * @ingroup surfaces
- *
- * @brief identify neighbours of a surface patch
- *
- * @param s an ::nbi_surface_t;
- * @param p index of patch whose neighbours are to be found;
- * @param r radius of neighbourhood;
- * @param nbrs on exit contains indices of nodes of \a s lying with
- * distance \a r of centroid of patch \a p of \a s;
- * @param nnbrs on exit contains number of entries in \a nbrs;
[…]
