### 2024-01-21 85433c4
maint: checking for PETSc using pkg-config
