+++ config.h
@@ -48,7 +48,7 @@
 /* Define to 1 if you have the <memory.h> header file. */
 #define HAVE_MEMORY_H 1
 
-/* PETSC solvers available */
+/* Have PETSc */
 #define HAVE_PETSC 1
 
 /* singular quadrature library available */
