### 2024-01-21 cf2b9d9
doc: updated to README.md covering installation of PETSc
