### 2024-01-23 15847d7
maint: exit from solvers if KSP file not found
