### 2024-01-26 7a2ac80
maint: updated interface to user-defined matrix multiplication in PETSc code
