### 2024-01-31 #5972 @mjcarley
I have gone through an installation on an otherwise absolutely clean cloud-based system. Installation is carried out using a script containing (the environment variables used reflect the details of installation in my personal account under .../Codes/... ):

rm -rf nbi blaswrap
git clone https://github.com/mjcarley/blaswrap
git clone --recursive https://github.com/mjcarley/nbi
cd blaswrap
. autogen.sh
./configure --prefix=$HOME/Codes --with-blas=openblas \
CPPFLAGS=-I$HOME/Codes/include LDFLAGS=-L$HOME/Codes/lib
make
make install
cd ..
cd nbi
. autogen.sh
./configure --prefix=$HOME/Codes \
CPPFLAGS=-I$HOME/Codes/include LDFLAGS="-L$HOME/Codes/lib -L$HOME/Codes/lib64"
make
