### 2024-01-31 #5972 @mjcarley
Release information says:

CentOS Linux release 7.8.2003 (Core)
