### 2024-03-14 #5972 @mscroggs
Sorry for being really slow on this. Some comments from me:

- **State of the field**: I'd like to see a little more in the paper saying what NBI offers that existing software does not.
- **Substantial scholarly effort**: Once you've added to **State of the field**, I think this will be clearer
- **Statement of need (documentation)**: There currently isn't a statement of need in the repo documentation. I recommend copying the into of the paper into the README.
- **Installation**: Following the install instructions told me to install https://github.com/mjcarley/blaswrap. When doing `make` in blaswrap, I get some cryptic errors about libopenblas. If you were able to provide a Docker image or similar with the dependencies and software installed, I think this would be very appreciated by users.
- **Automated tests**: There are some instructions to run tests in the examples folder, which I'll try running after I've managed to install. It would be great if you could get GitHub Actions to automatically run these.
- **Community guidelines**: I couldn't find this. You're welcome to borrow from [CONTRIBUTING.md from my software Symfem](https://github.com/mscroggs/symfem/blob/main/CONTRIBUTING.md). I'd also recommend adding [the Contributor Covenant code of conduct for contributors](https://www.contributor-covenant.org/version/2/0/code_of_conduct/).
- **References**: Looks like a couple DOIs are missing, otherwise these look fine
