### 2024-03-21 d5fcd50
doc: updated JOSS paper; added CONTRIBUTING.md and Code of Conduct
