+++ JOSS/paper.bib
@@ -118,7 +118,8 @@ doi = {10.1016/j.jcpx.2021.100092}
   journal = 	 {Computers and Mathematics with Applications},
   year = 	 2003,
   volume =	 45,
-  pages =	 {1829--1840}
+  pages =	 {1829--1840},
+  doi =          {10.1016/S0898-1221(03)90004-6}
 }
 
 @Article{xiao-gimbutas10,
