+++ CONTRIBUTING.md
@@ -1,0 +1,17 @@
+## How to contribute
+
+### Reporting bugs
+If you find a bug in NBI, please report it on the [issue tracker](https://github.com/mjcarley/nbi/issues/new?assignees=&labels=bug&template=bug_report.md&title=).
+
+### Suggesting enhancements
+If you want to suggest a new feature or an improvement of a current feature, you can submit this
+on the [issue tracker](https://github.com/mjcarley/nbi/issues/new?assignees=&labels=enhancement&template=feature_request.md&title=).
+
+### Submitting a pull request
+If you want to directly submit code to NBI, you can do this by forking the NBI repo, then submitting a pull request.
+If you want to contribute, but are unsure where to start, have a look at the
+[issues labelled "good first issue"](https://github.com/mjcarley/nbi/issues?q=is%3Aopen+is%3Aissue+label%3A%22good+first+issue%22).
+
+### Code of conduct
+We expect all our contributors to follow [the Contributor Covenant](CODE_OF_CONDUCT.md). Any unacceptable
+behaviour can be reported to Michael (codes@paraffinalia.co.uk)
