### 2024-03-26 #5972 @mjcarley
I've added the community guidelines and code of conduct.

The references that have a doi now have them in the reference list.

Can you say what the "cryptic errors" were when you installed blaswrap and nbi?
