### 2024-03-28 #5972 @bonh
If I try to install blaswrap I get: undefined reference to `dtrtrs_'
