### 2024-04-11 ab71444
feat: updating autoools files for compatibility with Docker image
