### 2024-04-11 dafa811
feat: updating autoools files for compatibility with Docker image
