### 2024-04-11 248e80d
feat: updating autoools files for compatibility with Docker image
