### 2024-04-11 f909b88
main: check for GMSH version in initialization for compatibility with Ubuntu Docker image
