### 2024-04-23 e637fff
maint: added library option for triangle-api when agg is being compiled in from module
