### 2024-04-23 c6ee45f
maint: added library option for triangle-api when agg is being compiled in from module
