### 2024-04-23 863ea18
maint: static link to triangle for when NBI is being compiled with AGG in Docker
