### 2024-04-23 de3730f
doc: added information on geometry generation to paper.md
