+++ JOSS/paper.md
@@ -59,6 +59,12 @@ allows boundary conditions to be specified in a form which makes it
 easier to perform parametric studies by systematically modifying
 internal variables. 
 
+Simple geometries, such as spheres and ellipsoids, are implemented as
+internal functions. More general shapes can be produced using standard
+GMSH inputs [@geuzaine-remacle09] or the CST method of Kulfan
+[@kulfan10] which is implemented in an extended and more general form
+in the AGG library which is included with the NBI distribution.
+
 # Mathematics
 
 The Laplace and Helmholtz potentials generated by a surface source
