+++ JOSS/paper.bib
@@ -105,6 +105,19 @@ doi = {10.1016/j.jcpx.2021.100092}
                   finite elem ent method, open-source software}
 }
 
+@article{kulfan10,
+author={Kulfan, B. M.},
+title = {Recent extensions and applications of the `{CST}' universal
+         parametric geometry representation method},
+volume = {114},
+doi = {10.1017/S0001924000003614},
+number = {1153},
+journal=aero,
+publisher={Cambridge University Press},
+year={2010},
+pages={157--176}
+}
+
 @Book{kirkup07,
   author = 	 {S. M. Kirkup},
   title = 	 {The {Boundary} {Element} {Method} in acoustics},
