### 2024-04-23 ad8a2cb
doc: added Dockerfile
