+++ JOSS/paper.bib
@@ -35,7 +35,7 @@ doi = {10.1016/j.jcpx.2021.100092}
 }
 
 @Book{gumerov-duraiswami04,
-  author = 	 {N. A. Gumerov and R. Duraiswami},
+  author = 	 {N. A Gumerov and R. Duraiswami},
   title = 	 {Fast Multipole Methods for the {Helmholtz} equation
                   in three dimensions},
   publisher = 	 {Elsevier},
