### 2024-04-23 fe1b724
Update docker-publish.yml removed cron scheduling
