+++ .github/workflows/docker-publish.yml
@@ -6,8 +6,8 @@ name: Docker
 # documentation.
 
 on:
-  schedule:
-    - cron: '45 16 * * *'
+  #schedule:
+  #  - cron: '45 16 * * *'
   push:
     branches: [ "main" ]
     # Publish semver tags as releases.
@@ -43,7 +43,9 @@ jobs:
         if: github.event_name != 'pull_request'
         uses: sigstore/cosign-installer@6e04d228eb30da1757ee4e1dd75a0ec73a653e06 #v3.1.1
         with:
-          cosign-release: 'v2.1.1'
+          cosign-release: 'v2.2.3'
+      # line above changed version according to 
+      # https://github.com/sigstore/cosign/issues/3614
 
       # Set up BuildKit Docker container builder to be able to build
       # multi-platform images and export cache
