### 2024-04-23 237eb27
Update docker-publish.yml with cosign 2.2.4
