+++ .github/workflows/docker-publish.yml
@@ -41,11 +41,13 @@ jobs:
       # https://github.com/sigstore/cosign-installer
       - name: Install cosign
         if: github.event_name != 'pull_request'
-        uses: sigstore/cosign-installer@6e04d228eb30da1757ee4e1dd75a0ec73a653e06 #v3.1.1
+        uses: sigstore/cosign-installer@v3.5.0
         with:
-          cosign-release: 'v2.2.3'
+          cosign-release: 'v2.2.4'
       # line above changed version according to 
-      # https://github.com/sigstore/cosign/issues/3614
+      # https://github.com/sigstore/cosign/issues/3614 
+      # and
+      # https://github.com/sigstore/cosign-installer
 
       # Set up BuildKit Docker container builder to be able to build
       # multi-platform images and export cache
