### 2024-04-24 e06f11c
Update docker-publish.yml with "latest" tag
