### 2024-04-24 b634482
Update docker-publish.yml with "latest" tag
