### 2024-04-24 2627418
Update docker-publish.yml to force complete rebuild
