### 2024-04-24 48392ff
Update docker-publish.yml to force complete rebuild
