### 2024-04-24 0656364
feat: rewrote example scripts to better adapt to Docker installation
