### 2024-04-24 9ce7262
feat: added tests to run under Docker
