### 2024-04-24 174535a
doc: updated README.md to include instructions on using Docker image
