+++ .github/workflows/docker-publish.yml
@@ -2,7 +2,8 @@
 name: Create and publish a Docker image
 
 # Configures this workflow to run every time a change is pushed to the branch called `release`.
-on: [push]
+on: [workflow_dispatch]
+##on: [push]
   #push:
    # branches: ['release']
 
