### 2024-04-24 d8b65e7
doc: expanded paper.md with information on features of code
