### 2024-04-24 bcc0209
doc: expanded paper.md with information on features of code
