### 2024-04-25 #5972 @vissarion
> Can you say what the "cryptic errors" were when you installed blaswrap and nbi?

@mscroggs Could you please reply to this question?
