### 2024-04-25 #5972 @mjcarley
I have taken the suggestion of building a Docker image which is now available on the github repository. The image is built with two of the test cases run, and the other tests can be run in a shell in the Docker container. There is a corresponding update to the README.md, with information about running the examples, or other computations.

I have also updated the statement of features of the code to note the use of the CST approach of Kulfan to generate quite general "aerodynamic" geometries.
