### 2024-04-25 #5972 @mscroggs
> Can you say what the "cryptic errors" were when you installed blaswrap and nbi?

Sorry completely missed this question when you posted it. I think the useful line of the long error is:

```/usr/bin/ld: blaswrap-tests.o: undefined reference to symbol 'dtrtrs_'```
