### 2024-04-25 5904f12
Drop files from .gitignore
