### 2024-05-27 #5972 @vissarion
@danielskatz thanks! There are several parts in this review that needs an update.

1. There is a problem for one reviewer (@mscroggs) in installing the package. See https://github.com/openjournals/joss-reviews/issues/5972#issuecomment-2076766108 that remains unaddressed. @mjcarley could you please reply?
2. The are two issues with a proposed solution by the author https://github.com/mjcarley/nbi/issues/2 and https://github.com/mjcarley/nbi/issues/5 but the reviewer (@bonh) didn't reply yet. @bohn could you please update the issue and if you find it fixed could you please close it as completed? @bonh is there something else that holds you back?
