### 2024-05-28 #5972 @vissarion
@mscroggs are you still getting the installation error? Did you manage to use the docker image to install the package?
