### 2024-06-17 13f8669
bug: fixed return code in nbi-config.in
