### 2024-06-19 #5972 @mscroggs
Sorry, I was on leave with little internet for 2 weeks then away at a conference. I was able to run the software in the docker image (thanks for providing this). Updating my checklist now
