### 2024-06-19 #5972 @vissarion
@bonh could you please update us on your review progress? Last time you had issues in compiling the source code. Did you manage to try the docker image? That approach was successful for @mscroggs who was also experiencing installation issues.
