### 2024-07-17 bfbd744
fact: changed gchar to char, as recommended by GLIB-2.0 documentation
