+++ tools/nbi-solve-helmholtz.c
@@ -52,7 +52,7 @@ PetscErrorCode nbi_petsc_MatMult_complex(Mat mat, Vec x, Vec y) ;
 #include "nbi-private.h"
 
 GTimer *timer ;
-gchar *progname ;
+char *progname ;
 
 /**
  * @page nbisolvehelmholtz Solving Helmholtz problems
@@ -149,7 +149,7 @@ static void print_help_text(FILE *f, gint depth,
   return ;
 }
 
-static gboolean check_file_read(gchar *file)
+static gboolean check_file_read(char *file)
 
 {
   FILE *f ;
@@ -161,7 +161,7 @@ static gboolean check_file_read(gchar *file)
   return TRUE ;
 }
 
-gint main(gint argc, gchar **argv)
+gint main(gint argc, char **argv)
 
 {
   nbi_surface_t *s ;
@@ -170,7 +170,7 @@ gint main(gint argc, gchar **argv)
   gdouble *f, *xp, *src, t, emax, fmax, e2, f2, *G, dtree, pwt, nwt, erms ;
   gdouble xfield[3], *p, *rhs ;
   FILE *output, *input ;
-  gchar ch, *gfile, *mfile, *bfile, *sfile, *kspfile ;
+  char ch, *gfile, *mfile, *bfile, *sfile, *kspfile ;
   gdouble *work, tol, k, hyp, h1, h2, error ;
   gint fmm_work_size, nqfmm, order_fmm, order_inc, i, fstr, solver_work_size ;
   gint gmres_max_iter, gmres_restart ;
@@ -188,7 +188,7 @@ gint main(gint argc, gchar **argv)
   PetscMPIInt size ;
   gint psize ;
   gpointer petsc_ctx[NBI_SOLVER_DATA_SIZE] ;
-  gchar *help = "" ;
+  char *help = "" ;
 #endif /*HAVE_PETSC*/
 
   nthreads = 1 ;
@@ -576,7 +576,7 @@ gint main(gint argc, gchar **argv)
       }
       PetscCall(PetscInitialize(&i, &argv, kspfile, help)) ;
     } else {
-      PetscCall(PetscInitialize(&i, &argv, (gchar *)0, help)) ;
+      PetscCall(PetscInitialize(&i, &argv, (char *)0, help)) ;
     }
     
     PetscCallMPI(MPI_Comm_size(PETSC_COMM_WORLD, &size));
