### 2024-07-25 37c9cd2
chore: update version number for release
