+++ config.h
@@ -94,7 +94,7 @@
 #define NBI_MICRO_VERSION 0
 
 /* Minor version */
-#define NBI_MINOR_VERSION 0
+#define NBI_MINOR_VERSION 1
 
 /* Name of package */
 #define PACKAGE "nbi"
@@ -106,7 +106,7 @@
 #define PACKAGE_NAME "nbi"
 
 /* Define to the full name and version of this package. */
-#define PACKAGE_STRING "nbi 1.0.0"
+#define PACKAGE_STRING "nbi 1.1.0"
 
 /* Define to the one symbol short name of this package. */
 #define PACKAGE_TARNAME "nbi"
@@ -115,10 +115,10 @@
 #define PACKAGE_URL ""
 
 /* Define to the version of this package. */
-#define PACKAGE_VERSION "1.0.0"
+#define PACKAGE_VERSION "1.1.0"
 
 /* Define to 1 if you have the ANSI C header files. */
 #define STDC_HEADERS 1
 
 /* Version number of package */
-#define VERSION "1.0.0"
+#define VERSION "1.1.0"
