+++ src/documentation.c
@@ -1,6 +1,6 @@
 /*
  * Autogenerated file, do not  edit
- * Thu Apr 25 10:36:22 BST 2024
+ * Wed Jul 17 12:18:05 BST 2024
  * -------------------------------
  */
 
@@ -437,7 +437,7 @@ gint nbi_mesh_triangulate(nbi_surface_t *s,
  * @return 0 on success.
  */
 
-gint nbi_mesh_export_gmsh(FILE *f, gchar *view,
+gint nbi_mesh_export_gmsh(FILE *f, char *view,
 				gdouble *x, gint xstr, gint np,
 				gint offp,
 				gint *tri, gint tstr, gint nt,
