### 2024-07-25 #5972 @mjcarley
The new release has been archived as v1.1.0 and the with DOI:10.5281/zenodo.12819049
