+++ JOSS/paper.bib
@@ -120,7 +120,7 @@ pages={157--176}
 
 @Book{kirkup07,
   author = 	 {S. M. Kirkup},
-  title = 	 {The {Boundary} {Element} {Method} in acoustics},
+  title = 	 {The Boundary Element Method in acoustics},
   publisher = 	 {Integrated Sound Software},
   year = 	 2007,
   url = {http://www.boundary-element-method.com}}
