### 2024-07-25 fdf3781
minor changes paper.bib
