+++ JOSS/paper.bib
@@ -45,7 +45,7 @@ doi = {10.1016/j.jcpx.2021.100092}
   }
 
 @TechReport{gumerov-duraiswami05,
-  author = 	 {Nail A. Gumerov and Ramani Duraiswami},
+  author = 	 {N. A. Gumerov and R. Duraiswami},
   title = 	 {Comparison of the efficiency of translation
                   operators used in the fast multipole method for the
                   {3D} {Laplace} equation}, 
@@ -54,7 +54,7 @@ doi = {10.1016/j.jcpx.2021.100092}
   number = 	 {CSTR~4701, UMIACS TR-2005-09}}
 
 @Article{gumerov-duraiswami03,
-  author = 	 {Nail A. Gumerov and Ramani Duraiswami},
+  author = 	 {N. A. Gumerov and R. Duraiswami},
   title = 	 {Recursions for the computation of multipole
                   translation and rotation coefficients for the {3-D}
                   {Helmholtz} equation},
@@ -66,7 +66,7 @@ doi = {10.1016/j.jcpx.2021.100092}
   doi =          {10.1137/S1064827501399705}}
 
 @Article{gumerov-duraiswami09,
-  author = 	 {Nail A. Gumerov and Ramani Duraiswami},
+  author = 	 {N. A. Gumerov and R. Duraiswami},
   title = 	 {A broadband fast multipole accelerated boundary
                   element method for the three dimensional {Helmholtz}
                   equation}, 
