### 2024-07-25 #5972 @danielskatz
@mjcarley - Can you also merge https://github.com/mjcarley/nbi/pull/7 - I think this should clear up the references of all four Gumerov ... papers which look a little odd in the PDF now as a group of 3 and 1.
