### 2024-07-25 #5972 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
