### 2024-07-25 #5972 @danielskatz
@openjournals/dev - Do you understand why these four references are showing up grouped into 3 and 1, rather than 4 together?
