### 2024-07-29 #5972 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
