### 2024-07-29 #5972 @danielskatz
Congratulations to @mjcarley (Michael Carley) on your publication!!

JOSS wouldn't be able to succeed without your voluntary effort
